Add files via upload
This code contains the most recent edit as of 04-27-2018 3:32PM. It compiles, but it might not post correctly to Firebase, as that is yet to be tested. The webhook calls are in mainApp.ino under the methods "getDataHandler" which reads from Firebase and "publishData" which writes to Firebase. The loop() writes/reads from Firebase every 20 seconds or so in real time.